Update Guilds, ServerLauncher, Timing/Misc Dependencies
Guilds should now be processing update requests on a global scale now. The new Guild System fetches guild updates from the database and processing them in batches. This will allow for cross server guild updates(add/remove/change rank/etc).

Server Launcher has been updated with a BootList. You can specify which scenes should be additively loaded if no launch parameters are found. The ClientBootstrap, LoginServer, WorldServer, and SceneServer scenes can now run simultaneously in the editor (with bugs.. due to the TimeManager) when starting the ServerLauncher scene.

Database services have been categorized into specific folders for clarity.

Fixed an issue with Billboards throwing null reference exception.

+++ FishMMO/Assets/Scripts/Server/World/WorldServer/WorldSceneSystem.cs
@@ -64,9 +64,12 @@ namespace FishMMO.Server
 
 		private void OnApplicationQuit()
 		{
-			using var dbContext = Server.DbContextFactory.CreateDbContext();
-			PendingSceneService.Delete(dbContext, Server.WorldServerSystem.ID);
-			dbContext.SaveChanges();
+			if (Server != null && Server.DbContextFactory != null)
+			{
+				using var dbContext = Server.DbContextFactory.CreateDbContext();
+				PendingSceneService.Delete(dbContext, Server.WorldServerSystem.ID);
+				dbContext.SaveChanges();
+			}
 		}
 
 		private void LateUpdate()
@@ -75,12 +78,15 @@ namespace FishMMO.Server
 			{
 				nextWaitQueueUpdate = waitQueueRate;
 
-				using var dbContext = Server.DbContextFactory.CreateDbContext();
-				foreach (string sceneName in new List<string>(WaitingConnections.Keys))
+				if (Server != null && Server.DbContextFactory != null)
 				{
-					TryClearWaitQueues(dbContext, sceneName);
+					using var dbContext = Server.DbContextFactory.CreateDbContext();
+					foreach (string sceneName in new List<string>(WaitingConnections.Keys))
+					{
+						TryClearWaitQueues(dbContext, sceneName);
+					}
+					dbContext.SaveChanges();
 				}
-				dbContext.SaveChanges();
 			}
 			nextWaitQueueUpdate -= Time.deltaTime;
 		}
